TST: move tests and setup CI (#13)
