+++ tests/arlington_test.py
@@ -4,9 +4,12 @@ import httpretty
 import requests
 import re
 from datetime import datetime
-from lambda_function import get_arlington_events, html_textraction, parse_event_name, \
+import sys
+from os import path
+sys.path.append( path.dirname( path.dirname( path.abspath(__file__) ) ) )
+from lambdas.arlington.lambda_function import get_arlington_events, html_textraction, parse_event_name, \
                             schematize_events, schematize_date
-from test_fixtures import page_one_uri_json, page_two_uri_json, page_one_uri_event_items, \
+from fixtures.arlington_test_fixtures import page_one_uri_json, page_two_uri_json, page_one_uri_event_items, \
                           page_two_uri_event_items, schematized_page_two_event_items
 
 url_regex = re.compile(
