+++ tests/fairfax_test.py
@@ -4,11 +4,14 @@ import httpretty
 import requests
 import re
 from datetime import datetime
-from lambda_function import get_event_cost, get_event_date_from_event_website, \
+import sys
+from os import path
+sys.path.append( path.dirname( path.dirname( path.abspath(__file__) ) ) )
+from lambdas.fairfax.lambda_function import get_event_cost, get_event_date_from_event_website, \
                             get_event_start_date, get_start_times, get_event_description, \
                             get_event_venue, parse_event_website, schematize_event_date, \
                             schematize_event_time, get_fairfax_events
-from test_fixtures import get_event_page_soup, get_calendar_page_soup, get_start_times_expected, \
+from fixtures.fairfax_test_fixtures import get_event_page_soup, get_calendar_page_soup, get_start_times_expected, \
                           canceled_page_content, get_fairfax_events_expected
 
 url_regex = re.compile(
