+++ tests/nps_test.py
@@ -4,9 +4,12 @@ import httpretty
 import re
 from datetime import datetime
 import requests
-from lambda_function import get_park_events, get_nps_events, get_specific_event_location, \
+import sys
+from os import path
+sys.path.append( path.dirname( path.dirname( path.abspath(__file__) ) ) )
+from lambdas.nps.lambda_function import get_park_events, get_nps_events, get_specific_event_location, \
                             schematize_nps_event, schematize_time, main
-from test_fixtures import get_park_events_expected, nama_events_json, event_page_content, \
+from fixtures.nps_test_fixtures import get_park_events_expected, nama_events_json, event_page_content, \
                           schematize_nps_event_expected
 
 url_regex = re.compile(
